Extract spi module for identity hub store and use EDC defaulting mechanism for in memory implementation
