+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/processor/CollectionsQueryProcessor.java
@@ -1,8 +1,8 @@
 package org.eclipse.dataspaceconnector.identityhub.processor;
 
-import org.eclipse.dataspaceconnector.identityhub.dtos.HubObject;
 import org.eclipse.dataspaceconnector.identityhub.dtos.MessageResponseObject;
 import org.eclipse.dataspaceconnector.identityhub.dtos.MessageStatus;
+import org.eclipse.dataspaceconnector.identityhub.store.HubObject;
 import org.eclipse.dataspaceconnector.identityhub.store.IdentityHubStore;
 
 import java.util.ArrayList;
@@ -15,7 +15,7 @@ import static org.eclipse.dataspaceconnector.identityhub.dtos.MessageResponseObj
  */
 public class CollectionsQueryProcessor implements MessageProcessor {
 
-    private IdentityHubStore identityHubStore;
+    private final IdentityHubStore identityHubStore;
 
     public CollectionsQueryProcessor(IdentityHubStore identityHubStore) {
         this.identityHubStore = identityHubStore;
