+++ extensions/identity-hub/src/main/java/org/eclipse/dataspaceconnector/identityhub/store/IdentityHubInMemoryStore.java
@@ -1,7 +1,5 @@
 package org.eclipse.dataspaceconnector.identityhub.store;
 
-import org.eclipse.dataspaceconnector.identityhub.dtos.HubObject;
-
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
