+++ extensions/identity-hub/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -1,2 +1,1 @@
-org.eclipse.dataspaceconnector.identityhub.api.IdentityHubExtension
-org.eclipse.dataspaceconnector.identityhub.store.IdentityHubInMemoryStoreExtension
+org.eclipse.dataspaceconnector.identityhub.IdentityHubExtension
