+++ extensions/identity-hub/src/test/java/org/eclipse/dataspaceconnector/identityhub/api/IdentityHubControllerTest.java
@@ -1,4 +1,4 @@
-package org.eclipse.dataspaceconnector.identityhub.api.featuredetection;
+package org.eclipse.dataspaceconnector.identityhub.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.javafaker.Faker;
