+++ settings.gradle.kts
@@ -1,4 +1,5 @@
 rootProject.name = "identity-hub"
 
 include(":extensions:identity-hub")
+include(":extensions:identity-hub-store-spi")
 include(":rest-client")
